requirements.txt
matplotlib
peakutils
pillow